requirements.txt
imageio==2.9.0
itk==5.2.0.post2
itk-core==5.2.0.post2
itk-elastix==0.12.0
itk-filtering==5.2.0.post2
itk-io==5.2.0.post2
itk-meshtopolydata==0.7.1
itk-numerics==5.2.0.post2
itk-registration==5.2.0.post2
itk-segmentation==5.2.0.post2
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
pydicom==2.1.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
